+++ fixtures/packages/polymer/expected/test/unit/dynamic-imports/dynamic-element.js
@@ -1,4 +1,4 @@
-import '../../../polymer.js';
+import '../../../polymer-legacy.js';
 import { Polymer } from '../../../lib/legacy/polymer-fn.js';
 import { html } from '../../../lib/utils/html-tag.js';
 import { dom } from '../../../lib/legacy/polymer.dom.js';