+++ fixtures/packages/polymer/expected/test/unit/events-elements.js
@@ -1,4 +1,4 @@
-import { Base } from '../../polymer.js';
+import { Base } from '../../polymer-legacy.js';
 import { Polymer } from '../../lib/legacy/polymer-fn.js';
 import { html } from '../../lib/utils/html-tag.js';
 /**